backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.3.1
firebase-admin==6.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
yfinance==0.2.28
pandas==2.0.3
python-dateutil==2.8.2
